+++ katarl2/agents/__init__.py
@@ -4,5 +4,3 @@ from .simba_sac.simba_sac import SimbaSAC
 from .simba_sac.simba_sac_cfg import SimbaSACConfig
 from .ppo.ppo import PPO
 from .ppo.ppo_cfg import PPODiscreteConfig, PPOContinuousConfig, SimbaPPOContinuousConfig, SimbaPPODiscreteConfig
-from .simba_ppo.simba_ppo import SimbaPPO
-from .simba_ppo.simba_ppo_cfg import SimbaPPOConfig