+++ CHANGELOG.md
@@ -52,6 +52,8 @@ P.S.
 ## v0.4 (20250903 -)
 1. 由于无法保证envpool和gymnasium的实验一致性, 删除了envpoolq全部环境, 最后存在版本为v0.3 9004d097a461996f19365fc3287acbb50c614624
 2. 统一max_and_skip为action_repeat, 并加入action_repeat_wrapper来是否用action_repeat_wrppaer或atari_wrappers
+3. 加入对num_train_steps的保存, 用于resume模型后, 继续训练能保存正确的模型名称
+4. 删除单独的SimbaPPO算法, 已完全整合在PPO算法中
 
 TODO:
 - [ ] 加入DreamerV3